refactor client admin container structure
