+++ app/client/container/admin/analytics/index.tsx
@@ -1,0 +1,1 @@
+export { default as Progress } from './Progress';
