+++ app/client/container/admin/groups/group.tsx
@@ -15,15 +15,17 @@ import ContentAdd from 'material-ui/svg-icons/content/add';
 import Divider from 'material-ui/Divider';
 import { Tabs, Tab } from 'material-ui/Tabs';
 
-import CreateGroupDialog from './create_group_dialog';
+import GroupUsers from './group_users';
+import GroupCollections from './group_collections';
 
 // selectors
 import { select_group } from 'client/state/groups/selectors';
 import { get_users_by_group } from 'client/state/users/selectors';
+import { select_collections_by_ids } from 'client/state/collections/selectors';
 
 // types
 import { IState } from 'client/state';
-import { IGroup, IUser } from 'common/types';
+import { IGroup, IUser, ICollection } from 'common/types';
 
 // actions
 import {
@@ -34,8 +36,10 @@ import {
 
 interface IStateProps {
     group: IGroup;
+    collections: ICollection[];
     users: IUser[];
     group_id: string;
+    tab: string;
 }
 
 interface IDispatchProps {
@@ -69,33 +73,52 @@ export class AdminGroup extends React.Component<IProps, {}> {
                         zIndex: 1099,
                         width: '100%'
                     }}
+                    tabItemContainerStyle={{
+                        background: 'linear-gradient(120deg, #8e44ad, #3498db)'
+                    }}
+                    value={this.props.tab}
                 >
-                    <Tab label="Settings">
+                    <Tab
+                        label="Settings"
+                        value="settings"
+                        onActive={() =>
+                            this.props.dispatch(
+                                push(
+                                    '/admin/groups/' +
+                                        this.props.group_id +
+                                        '/settings'
+                                )
+                            )}
+                    >
                         <div>test</div>
                     </Tab>
-                    <Tab label="Users">
-                        <List>
-                            {this.props.users.map(user => (
-                                <div>
-                                    <ListItem
-                                        leftAvatar={
-                                            <Avatar>
-                                                {user.name.substring(0, 3)}
-                                            </Avatar>
-                                        }
-                                        primaryText={user.name}
-                                        onClick={() =>
-                                            this.props.dispatch(
-                                                push('/admin/users/' + user._id)
-                                            )}
-                                    />
-                                    <Divider inset={true} />
-                                </div>
-                            ))}
-                        </List>
+                    <Tab
+                        label="Users"
+                        value="users"
+                        onActive={() =>
+                            this.props.dispatch(
+                                push(
+                                    '/admin/groups/' +
+                                        this.props.group_id +
+                                        '/users'
+                                )
+                            )}
+                    >
+                        <GroupUsers {...this.props} />
                     </Tab>
-                    <Tab label="Assignments">
-                        <div>test2</div>
+                    <Tab
+                        label="Collections"
+                        value="collections"
+                        onActive={() =>
+                            this.props.dispatch(
+                                push(
+                                    '/admin/groups/' +
+                                        this.props.group_id +
+                                        '/collections'
+                                )
+                            )}
+                    >
+                        <GroupCollections {...this.props} />
                     </Tab>
                 </Tabs>
             </div>
@@ -107,6 +130,11 @@ function mapStateToProps(state: IState, ownProps): IStateProps {
     return {
         group: select_group(state, ownProps.params.group_id),
         users: get_users_by_group(state, ownProps.params.group_id),
+        collections: select_collections_by_ids(
+            state,
+            select_group(state, ownProps.params.group_id).assigned_collections
+        ),
+        tab: ownProps.params.tab,
         group_id: ownProps.params.group_id
     };
 }
