+++ app/client/router.tsx
@@ -6,17 +6,17 @@ import Auth from 'client/container/auth';
 import Websocket from 'client/container/websocket';
 import Landing from 'client/container/landing';
 import AdminIndex from 'client/container/admin';
-import AdminCollections from 'client/container/admin/collections';
-import AdminGroups from 'client/container/admin/groups';
-import AdminGroup from 'client/container/admin/groups/group/group_index';
-import AdminGroupCreateOrAddUser from 'client/container/admin/groups/group/create_or_add_user_dialog';
-import AdminAddCollectionToGroup from 'client/container/admin/groups/group/add_collection_dialog';
-import AdminTags from 'client/container/admin/tags';
-import AdminCards from 'client/container/admin/cards';
-import AdminCard from 'client/container/admin/cards/card';
-import AdminUsers from 'client/container/admin/users';
-import AdminUser from 'client/container/admin/users/user';
-import AdminCollectionProgress from 'client/container/admin/analytics/progress';
+import { Collections } from 'client/container/admin/collections';
+import {
+    Group,
+    Groups,
+    CreateOrAddUserDialog,
+    AddCollectionDialog
+} from 'client/container/admin/groups';
+import { Tags } from 'client/container/admin/tags';
+import { Card, Cards } from 'client/container/admin/cards';
+import { User, Users } from 'client/container/admin/users';
+import { Progress } from 'client/container/admin/analytics';
 
 import UserIndex from 'client/container/user';
 import UserCollections from 'client/container/user/collections';
@@ -56,42 +56,33 @@ export default class RouterWrapper extends React.Component<IProps, {}> {
                             </Route>
                         </Route>
                         <Route path="/admin" component={AdminIndex}>
-                            <Route
-                                path="collections"
-                                component={AdminCollections}
-                            />
-                            <Route path="groups" component={AdminGroups} />
+                            <Route path="collections" component={Collections} />
+                            <Route path="groups" component={Groups} />
                             <Redirect
                                 from="groups/:group_id"
                                 to="groups/:group_id/users"
                             />
                             <Route
                                 path="groups/:group_id/:tab"
-                                component={AdminGroup}
+                                component={Group}
                             />
                             <Route
                                 path="groups/:group_id/users/add"
-                                component={AdminGroupCreateOrAddUser}
+                                component={CreateOrAddUserDialog}
                             />
                             <Route
                                 path="groups/:group_id/collections/add"
-                                component={AdminAddCollectionToGroup}
+                                component={AddCollectionDialog}
                             />
                             <Route
                                 path="analytics/progress"
-                                component={AdminCollectionProgress}
+                                component={Progress}
                             />
-                            <Route path="users" component={AdminUsers} />
-                            <Route
-                                path="users/:user_id"
-                                component={AdminUser}
-                            />
-                            <Route path="tags" component={AdminTags} />
-                            <Route path="cards" component={AdminCards} />
-                            <Route
-                                path="cards/:card_id"
-                                component={AdminCard}
-                            />
+                            <Route path="users" component={Users} />
+                            <Route path="users/:user_id" component={User} />
+                            <Route path="tags" component={Tags} />
+                            <Route path="cards" component={Cards} />
+                            <Route path="cards/:card_id" component={Card} />
                         </Route>
                     </Route>
                 </Route>
